refactor(core): moved all actions/intents to specific folders - updated imports to match
