+++ src/client/components/main/intent.js
@@ -5,11 +5,11 @@ import {mergeActionsByName} from '../../utils/actionUtils'
 export default function intent({DOM,socketIO}, other){
 
   //ugh, refactor
-  const feedActionsFromDOM = require('../../core/feeds/fromDOM').default(DOM)
-  const feedActionsFromSIO = require('../../core/feeds/fromSocketIO').default(socketIO)
+  const feedActionsFromDOM = require('../../core/feeds/actions/fromDOM').default(DOM)
+  const feedActionsFromSIO = require('../../core/feeds/actions/fromSocketIO').default(socketIO)
 
-  const nodeActionsFromDOM = require('../../core/nodes/fromDOM').default(DOM)
-  const nodeActionsFromSIO = require('../../core/nodes/fromSocketIO').default(socketIO)
+  const nodeActionsFromDOM = require('../../core/nodes/actions/fromDOM').default(DOM)
+  const nodeActionsFromSIO = require('../../core/nodes/actions/fromSocketIO').default(socketIO)
 
   const actions = mergeActionsByName([
     feedActionsFromDOM, feedActionsFromSIO,
