+++ src/client/core/feeds/actions/fromSocketIO.js
@@ -1,4 +1,4 @@
-import {toArray} from '../../utils/utils'
+import {toArray} from '../../../utils/utils'
 
 export default function actions(socketIO){
   const setFeedsData$ = socketIO.get("getFeedsData")
