+++ src/client/core/nodes/actions/fromSocketIO.js
@@ -1,4 +1,4 @@
-import {toArray} from '../../utils/utils'
+import {toArray} from '../../../utils/utils'
 
 export default function actions(socketIO){
   const setNodes$ =
